worker/requirements.txt
temporalio>=1.5.0
asyncio
python-dotenv>=1.0.0 